쓸모없는 파일 정리 가시고
