--- packages/core/src/lib/animator/index.ts
@@ -1,2 +1,0 @@
-//export { Animator } from "./legacy";
-export { Animator } from "./new";
